feat: 优化隐藏元素父级匹配策略和Tauri命令注册
✨ 核心改进：
- 完善 hidden_element_parent_strategy.rs 的策略算法和逻辑
- 优化 main.rs 中的 Tauri 命令注册和依赖管理
- 增强隐藏元素父级匹配的准确性和稳定性
- 提升后端匹配策略的性能和可靠性

🔧 技术优化：
- 重构隐藏元素策略的匹配算法
- 改进 Rust 后端的错误处理机制
- 优化 Tauri 命令的参数验证和类型安全
- 增强策略执行的日志记录和调试信息

🏗️ 架构改进：
- 完善匹配策略模块的组织结构
- 优化服务层的依赖注入和生命周期管理
- 增强策略接口的扩展性和可维护性
- 改进模块间的耦合度和代码复用性

📈 性能提升：
- 优化匹配算法的执行效率
- 减少不必要的计算开销
- 提升大数据量场景下的处理速度
- 改进内存使用和资源管理

+++ src-tauri/src/services/execution/matching/strategies/hidden_element_parent_strategy.rs
@@ -6,11 +6,10 @@
 use crate::services::execution::matching::strategies::{
     StrategyProcessor, MatchingContext, StrategyResult, ProcessingError
 };
-use crate::xml_judgment::{HiddenElementParentConfig};
+use crate::xml_judgment::HiddenElementParentConfig;
 use async_trait::async_trait;
 use anyhow::Result;
 use serde_json::Value;
-use std::collections::HashMap;
 
 pub struct HiddenElementParentStrategyProcessor;
 
@@ -111,7 +110,7 @@ impl HiddenElementParentStrategyProcessor {
         // 层级关系加分
         if self.is_potential_parent(candidate, hidden_element) { score += 0.1; }
         
-        score.min(1.0)
+        (score as f64).min(1.0)
     }
 
     /// 搜索父容器候选者
